# swagger.yaml
...
openapi: '3.1.0'
...
info:
  version: '0.0.1'
  title: Todo List API
servers:
  - url: http://localhost:4000/api
    description: Production server
schemes: -https
...
paths:
...
  /profile:
    get:
      tags:
        - profile
      description: get all profiles, requires api key
      produces:
        - application/json
      security:
        - apiKey: []
      responses:
        200:
          description: Get profile success
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/getAllProfilesSuccess'
        401:
          description: JWT expired or invalid
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/401-JWT-expired'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/500'
    post:
      tags:
        - profile
      description: add new profile, requires api key
      produces:
        - application/json
      security:
        - apiKey: []
      requestBody:
        description: add new profile
        content:
          application/json:
            schema:
              $ref: '#/components/schemes/add-profile'
      responses:
        201:
          description: Add profile success
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/addProfileSuccess'
        401:
          description: JWT expired or invalid
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/401-JWT-expired'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/500'
  /profile/{id}:
    get:
      tags:
        - profile
      description: find one profile by id, requires api key
      produces:
        - application/json
      security:
        - apiKey: []
      parameters:
        - name: id
          in: path
          description: id
          required: true
          explode: true
          schema:
            type: number
      responses:
        200:
          description: Get profile success
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/getProfileSuccess'
        400:
          description: Invalid id, requires format numeric
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/400-invalid-format-id'
        401:
          description: JWT expired or invalid
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/401-JWT-expired'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/500'
    put:
      tags:
        - profile
      description: update profile, requires api key
      produces:
        - application/json
      security:
        - apiKey: []
      parameters:
        - name: id
          in: path
          description: id
          required: true
          explode: true
          schema:
            type: string
      requestBody:
        description: update profile
        content:
          application/json:
            schema:
              $ref: '#/components/schemes/add-profile'
      responses:
        200:
          description: update profile success
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/updateProfilesSuccess'
        400:
          description: Invalid id, requires format numeric
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/400-invalid-format-id'
        401:
          description: JWT expired or invalid
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/401-JWT-expired'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/500'
    delete:
      tags:
        - profile
      description: delete profile by id, requires api key
      produces:
        - application/json
      security:
        - apiKey: []
      parameters:
        - name: id
          in: path
          description: id
          required: true
          explode: true
          schema:
            type: string
      responses:
        200:
          description: delete profile success
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/deleteProfilesSuccess'
        400:
          description: Invalid id, requires format numeric
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/400-invalid-format-id'
        401:
          description: JWT expired or invalid
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/401-JWT-expired'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/500'
...
  /todo:
    get:
      tags:
        - todo
      description: get all todo, requires api key
      produces:
        - application/json
      security:
        - apiKey: []
      responses:
        200:
          description: Get todo success
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/getAllTodoSuccess'
        401:
          description: JWT expired or invalid
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/401-JWT-expired'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/500'
    post:
      tags:
        - todo
      description: add new todo, requires api key
      produces:
        - application/json
      security:
        - apiKey: []
      requestBody:
        description: add new todo
        content:
          application/json:
            schema:
              $ref: '#/components/schemes/add-todo'
      responses:
        201:
          description: Add todo success
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/addTodoSuccess'
        401:
          description: JWT expired or invalid
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/401-JWT-expired'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/500'
  /todo/{id}:
    get:
      tags:
        - todo
      description: find one todo by id, requires api key
      produces:
        - application/json
      security:
        - apiKey: []
      parameters:
        - name: id
          in: path
          description: id
          required: true
          explode: true
          schema:
            type: number
      responses:
        200:
          description: Get todo success
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/getTodoSuccess'
        400:
          description: Invalid id, requires format numeric
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/400-invalid-format-id'
        401:
          description: JWT expired or invalid
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/401-JWT-expired'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/500'
    put:
      tags:
        - todo
      description: update todo, requires api key
      produces:
        - application/json
      security:
        - apiKey: []
      parameters:
        - name: id
          in: path
          description: id
          required: true
          explode: true
          schema:
            type: string
      requestBody:
        description: update todo
        content:
          application/json:
            schema:
              $ref: '#/components/schemes/add-todo'
      responses:
        200:
          description: update todo success
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/updateTodoSuccess'
        400:
          description: Invalid id, requires format numeric
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/400-invalid-format-id'
        401:
          description: JWT expired or invalid
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/401-JWT-expired'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/500'
    delete:
      tags:
        - todo
      description: delete todo by id, requires api key
      produces:
        - application/json
      security:
        - apiKey: []
      parameters:
        - name: id
          in: path
          description: id
          required: true
          explode: true
          schema:
            type: string
      responses:
        200:
          description: delete todo success
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/deleteTodoSuccess'
        400:
          description: Invalid id, requires format numeric
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/400-invalid-format-id'
        401:
          description: JWT expired or invalid
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/401-JWT-expired'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/500'
  /todo/frequency:
    post:
      tags:
        - todo
      description: add new todo by frequency, requires api key
      produces:
        - application/json
      security:
        - apiKey: []
      requestBody:
        description: add new todo by frequency
        content:
          application/json:
            schema:
              $ref: '#/components/schemes/add-todo-frequency'
      responses:
        201:
          description: Add todo success
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/addTodoFrequencySuccess'
        401:
          description: JWT expired or invalid
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/401-JWT-expired'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/500'
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      name: Authorization
      in: header
      format: JWT
  schemes:
...
    add-profile:
      type: object
      properties:
        name:
          type: string
        dob:
          type: string
        gender:
          type: string
        email:
          type: string
        image:
          type: string
      example:
        name: Xuan Long
        dob: 2002-11-19
        gender: male
        email: long@gmail.com
        image: adandwnajnda
...
    add-todo:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
        timeStart:
          type: string
        timeEnd:
          type: string
        account:
          id:
            type: number
      example:
        title: Hello
        description: hello
        status: todo
        timeStart: 2023-10-26T08:15:15.000Z
        timeEnd: 2023-10-27T08:15:15.000Z
        account:
          id: 5
    add-todo-frequency:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
        timeStart:
          type: string
        timeEnd:
          type: string
        frequency:
          type: string
        account:
          id:
            type: number
      example:
        title: Hello
        description: hello
        status: todo
        timeStart: 2023-10-26T08:15:15.000Z
        timeEnd: 2023-10-27T08:15:15.000Z
        frequency: 'weekly'
        account:
          id: 5
...
    getAllProfilesSuccess:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: number
              name:
                type: string
              dob:
                type: string
              gender:
                type: string
              email:
                type: string
              image:
                type: string
      example:
        data:
          - id: 1
            name: Xuan Long
            dob: 2002-11-19
            gender: male
            email: long@gmail.com
            image: adandwnajnda
    addProfileSuccess:
      type: object
      properties:
        message:
          type: string
        data:
          id:
            type: number
          name:
            type: string
          dob:
            type: string
          gender:
            type: string
          email:
            type: string
          image:
            type: string
      example:
        message: 'Profile id 13 created'
        data:
          id: 1
          name: Xuan Long
          dob: 2002-11-19
          gender: male
          email: long@gmail.com
          image: adandwnajnda
    getProfileSuccess:
      type: object
      properties:
        data:
          type: object
          properties:
            id:
              type: number
            name:
              type: string
            dob:
              type: string
            gender:
              type: string
            email:
              type: string
            image:
              type: string
      example:
        data:
          id: 1
          name: Xuan Long
          dob: 2002-11-19
          gender: male
          email: long@gmail.com
          image: adandwnajnda
    updateProfilesSuccess:
      type: object
      properties:
        message:
          type: string
        data:
          id:
            type: number
          name:
            type: string
          dob:
            type: string
          gender:
            type: string
          email:
            type: string
          image:
            type: string
      example:
        message: 'Profile id 1 updated successfully'
        data:
          id: 1
          name: Xuan Long
          dob: 2002-11-19
          gender: male
          email: long@gmail.com
          image: adandwnajnda
    deleteProfilesSuccess:
      type: object
      properties:
        message:
          type: string
      example:
        message: success
...
    getAllTodoSuccess:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: number
              title:
                type: string
              description:
                type: string
              status:
                type: string
              timeStart:
                type: string
              timeEnd:
                type: string
              account:
                id:
                  type: number
                username:
                  type: string
                password:
                  type: string
                role:
                  type: string
      example:
        data:
          - id: 19
            title: Hello
            description: hello
            status: todo
            timeStart: 2023-10-26T08:15:15.000Z
            timeEnd: 2023-10-27T08:15:15.000Z
            account:
              id: 5
              username: admin
              password: $2b$10$KZwBc0faUqgxrEca0bwqm.//1se4XUsqDrIsiuxk0VSqQ0yWhxSzS
              role: admin
    addTodoSuccess:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        description:
          type: string
        status:
          type: string
        timeStart:
          type: string
        timeEnd:
          type: string
        account:
          id:
            type: number
      example:
        data:
          account:
            id: 5
          title: Hello
          description: hello
          timeStart: 2023-10-26 15:15:15
          timeEnd: 2023-10-27 15:15:15
          status: todo
          id: 45
        message: todo id 45 created
    getTodoSuccess:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        description:
          type: string
        status:
          type: string
        timeStart:
          type: string
        timeEnd:
          type: string
        account:
          id:
            type: number
          username:
            type: string
          password:
            type: string
          role:
            type: string
      example:
        data:
          id: 19
          title: Hello
          description: hello
          status: todo
          timeStart: 2023-10-26T08:15:15.000Z
          timeEnd: 2023-10-27T08:15:15.000Z
          account:
            id: 5
            username: admin
            password: $2b$10$KZwBc0faUqgxrEca0bwqm.//1se4XUsqDrIsiuxk0VSqQ0yWhxSzS
            role: admin
    updateTodoSuccess:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        description:
          type: string
        status:
          type: string
        timeStart:
          type: string
        timeEnd:
          type: string
        account:
          id:
            type: number
      example:
        data:
          account:
            id: 5
          title: Hello
          description: hello
          timeStart: 2023-10-26 15:15:15
          timeEnd: 2023-10-27 15:15:15
          status: todo
          id: 45
        message: Success
    deleteTodoSuccess:
      type: object
      properties:
        message:
          type: string
      example:
        message: Todo deleted
    addTodoFrequencySuccess:
      message:
        type: string
      data:
        type: object
        properties:
          id:
            type: number
          title:
            type: string
          description:
            type: string
          status:
            type: string
          timeStart:
            type: string
          timeEnd:
            type: string
          account:
            id:
              type: number
      example:
        data:
          id: 45
          account:
            id: 5
          title: Hello
          description: hello
          timeStart: 2023-10-26 15:15:15
          timeEnd: 2023-10-27 15:15:15
          status: todo
        message: todo created
...
    400-invalid-format-id:
      type: object
      properties:
        error:
          type: string
      example:
        error: Validation failed (numeric string is expected)
...
    401-JWT-expired:
      type: object
      properties:
        error:
          type: string
      example:
        error: Invalid or expired JWT token
...
    500:
      type: object
      properties:
        error:
          type: string
      example:
        error: Internal server error
# security:
#   - apiKey: []